requirements.txt
pyppeteer==0.0.25
requests==2.21.0
Pillow==6.0.0
